Services/files.requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
python-magic
zipfile38
pillow
redis 
httpx